Inject executor id into ExecutionRequestResults (#4750)

+++ datahub-executor/src/datahub_executor/coordinator/server.py
@@ -1,13 +1,17 @@
 import logging
 import os
 import signal
-import time
-from threading import Thread
+from threading import Event, Thread
 
 from fastapi import FastAPI
 
+from datahub_executor.common.discovery.discovery import DatahubExecutorDiscovery
+from datahub_executor.common.helpers import create_datahub_graph
 from datahub_executor.common.monitoring.base import monitoring_start
-from datahub_executor.config import DATAHUB_EXECUTOR_GRACEFUL_SHUTDOWN_PERIOD
+from datahub_executor.config import (
+    DATAHUB_EXECUTOR_EMBEDDED_WORKER_ENABLED,
+    DATAHUB_EXECUTOR_GRACEFUL_SHUTDOWN_PERIOD,
+)
 from datahub_executor.coordinator.assertion_endpoints import assertions_router
 from datahub_executor.coordinator.health_endpoints import health_router
 from datahub_executor.coordinator.helpers import (
@@ -38,26 +42,47 @@ def shutdown_handler(*args, **kwargs):  # type: ignore
 
     logger.info("Shutdown handler: uvicorn initiated shutdown")
 
+    stop_event = Event()
+    stop_event.clear()
+    stop_succeeded = False
+
     # Give all jobs some time to finish, then forcefully terminate the process
     def force_shutdown() -> None:
-        time.sleep(DATAHUB_EXECUTOR_GRACEFUL_SHUTDOWN_PERIOD)
-        logger.info("Shutdown handler: graceful period expired, force-exiting...")
-        os.kill(os.getpid(), signal.SIGKILL)
+        stop_event.clear()
+        stop_event.wait(timeout=DATAHUB_EXECUTOR_GRACEFUL_SHUTDOWN_PERIOD)
+
+        if not stop_succeeded:
+            logger.info("Shutdown handler: graceful period expired, force-exiting...")
+            os.kill(os.getpid(), signal.SIGKILL)
 
     t = Thread(target=force_shutdown)
     t.start()
 
     for sighdlr in sighandler:
         try:
+            logger.error(f"shutting down: {sighdlr}")
             sighdlr()
         except Exception:
             logger.error(
                 "Shutdown handler: failed to execute one or more shutdown procedures"
             )
 
+    stop_succeeded = True
+    stop_event.set()
+    t.join()
+
+
+# Create graph instance
+graph = create_datahub_graph()
+
+# Start discovery process
+if DATAHUB_EXECUTOR_EMBEDDED_WORKER_ENABLED:
+    discovery = DatahubExecutorDiscovery(graph)
+    sighandler.append(discovery.stop)
+    discovery.start()
 
 # Start the scheduler
-start_scheduler(sighandler)
+start_scheduler(graph, sighandler)
 
 # Start the ingestion pipeline consumer
-start_ingestion_pipeline(sighandler)
+start_ingestion_pipeline(graph, discovery, sighandler)
